requirements.txt
# AiMind-OCR Docker 전용 requirements
# OCR 관련 의존성만 포함

# === FastAPI 및 기본 웹 프레임워크 ===
fastapi==0.123.9
uvicorn==0.40.0
python-multipart>=0.0.20
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
starlette==0.50.0

# === 이미지 처리 ===
pillow==12.1.0

# === OCR/diary_ocr_pipeline (VARCO, transformers) ===
transformers>=4.40.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
matplotlib>=3.7.0
torch>=2.0.0
torchvision>=0.15.0

# === Gemini API (후처리용) ===
google-generativeai==0.8.6
google-api-core==2.29.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2

# === 기본 유틸리티 ===
numpy>=1.26,<2
requests==2.32.5
httpx==0.28.1
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
